refactor 24.10.14 LottoRank 계산 Lotto에서 WinningLotto로 변경
