+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import static lotto.domain.LottoRank.SECOND;
-import static lotto.domain.LottoRank.THIRD;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -12,28 +10,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class LottoTest {
-    @DisplayName("로또등수")
-    @Test
-    void 로또등수() {
-        List<Integer> lottoNumbers = new ArrayList<>(Arrays.asList(1, 5, 7, 8, 9, 11));
-        Lotto lotto = new Lotto(lottoNumbers);
-        List<Integer> winningLottoNumbers = new ArrayList<>(Arrays.asList(1, 5, 7, 8, 9, 10));
-        int bonusNumber = 13;
-        WinningLotto winningLotto = new WinningLotto(new Lotto(winningLottoNumbers), bonusNumber);
-        assertThat(lotto.getLottoResult(winningLotto)).isEqualTo(THIRD);
-    }
-
-    @DisplayName("로또2등")
-    @Test
-    void 로또2등() {
-        List<Integer> lottoNumbers = new ArrayList<>(Arrays.asList(1, 5, 7, 8, 9, 11));
-        Lotto lotto = new Lotto(lottoNumbers);
-        List<Integer> winningLottoNumbers = new ArrayList<>(Arrays.asList(1, 5, 7, 9, 11, 12));
-        int bonusNumber = 8;
-        WinningLotto winningLotto = new WinningLotto(new Lotto(winningLottoNumbers), bonusNumber);
-        assertThat(lotto.getLottoResult(winningLotto)).isEqualTo(SECOND);
-    }
-
     @DisplayName("로또 발행시 로또 번호가 6개가 아닌 경우 에러")
     @Test
     void 로또_발행시_6개_아닐시_에러() {
